Rubocop実行後

+++ app/controllers/items_controller.rb
@@ -23,9 +23,7 @@ class ItemsController < ApplicationController
   end
 
   def edit
-    if current_user.id != @item.user_id || @item.order != nil
-      redirect_to root_path
-    end
+    redirect_to root_path if current_user.id != @item.user_id || !@item.order.nil?
   end
 
   def update
@@ -38,9 +36,7 @@ class ItemsController < ApplicationController
 
   def destroy
     item = Item.find(params[:id])
-    if current_user.id == item.user_id && item.order == nil
-      item.destroy
-    end
+    item.destroy if current_user.id == item.user_id && item.order.nil?
     redirect_to root_path
   end
